fix(core): Move interface class to API package
Move interface class to API package.

BREAKING CHANGE: Move interface class to API package.
