+++ jdklog-core/src/main/java/org/jdkstack/jdklog/logging/core/handler/ThreadMonitor.java
@@ -7,10 +7,10 @@ import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+import org.jdkstack.jdklog.logging.api.context.StudyThreadImpl;
 import org.jdkstack.jdklog.logging.api.context.WorkerContext;
 import org.jdkstack.jdklog.logging.api.metainfo.Constants;
 import org.jdkstack.jdklog.logging.api.monitor.Monitor;
-import org.jdkstack.jdklog.logging.core.context.StudyThreadImpl;
 
 /**
  * 定时检查线程的运行时间.
