+++ jdklog-core/src/main/java/org/jdkstack/jdklog/logging/core/manager/AbstractJuliLogger.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 import org.jdkstack.jdklog.logging.api.filter.Filter;
 import org.jdkstack.jdklog.logging.api.handler.Handler;
+import org.jdkstack.jdklog.logging.api.logger.ConfigurationDataImpl;
 import org.jdkstack.jdklog.logging.api.logger.Logger;
 import org.jdkstack.jdklog.logging.api.manager.LogManager;
 import org.jdkstack.jdklog.logging.api.metainfo.Level;
@@ -11,7 +12,7 @@ import org.jdkstack.jdklog.logging.api.metainfo.LogLevel;
 import org.jdkstack.jdklog.logging.api.metainfo.LogRecord;
 import org.jdkstack.jdklog.logging.api.metainfo.Record;
 import org.jdkstack.jdklog.logging.core.logger.ConfigurationData;
-import org.jdkstack.jdklog.logging.core.logger.ConfigurationDataImpl;
+
 
 /**
  * This is a class description.
